Fixed a comparison warning in 64-bit environments.

+++ src/vermouth/midvoice.c
@@ -95,7 +95,7 @@ static SAMPLE resample_vibrate(VOICE v, SAMPLE dst, SAMPLE dstterm) {
 		} while(dstbreak < dstterm);
 		v->sampstep = step;
 	}
-	v->vibrate.count = cnt - (dstterm - dst);
+	v->vibrate.count = cnt - (int)(dstterm - dst);
 	do {
 		RESAMPLING(dst, src, pos);
 		dst++;
@@ -147,7 +147,7 @@ static SAMPLE resample_vibloop(VOICE v, SAMPLE dst, SAMPLE dstterm) {
 		} while(dstbreak < dstterm);
 		v->sampstep = step;
 	}
-	v->vibrate.count = cnt - (dstterm - dst);
+	v->vibrate.count = cnt - (int)(dstterm - dst);
 	do {
 		RESAMPLING(dst, src, pos);
 		dst++;
@@ -212,7 +212,7 @@ rr_forward1:;
 			dstbreak += cnt;
 		} while(dstbreak < dstterm);
 	}
-	v->vibrate.count = cnt - (dstterm - dst);
+	v->vibrate.count = cnt - (int)(dstterm - dst);
 	do {
 		RESAMPLING(dst, src, pos);
 		dst++;
@@ -245,7 +245,7 @@ rr_backward1:;
 			dstbreak += cnt;
 		} while(dstbreak < dstterm);
 	}
-	v->vibrate.count = cnt - (dstterm - dst);
+	v->vibrate.count = cnt - (int)(dstterm - dst);
 	do {
 		RESAMPLING(dst, src, pos);
 		dst++;
@@ -525,7 +525,7 @@ static void mixrel_normal(VOICE v, SINT32 *dst, SAMPLE src, SAMPLE srcterm) {
 	SINT32	relr;
 	_SAMPLE	s;
 
-	samples = srcterm - src;
+	samples = (int)(srcterm - src);
 	voll = v->envleft << RELBIT;
 	rell = -(voll / samples);
 	if (!rell) {
@@ -557,7 +557,7 @@ static void mixrel_left(VOICE v, SINT32 *dst, SAMPLE src, SAMPLE srcterm) {
 	_SAMPLE	s;
 
 	vol = v->envleft << RELBIT;
-	rel = - (vol / (srcterm - src));
+	rel = - (vol / (int)(srcterm - src));
 	if (!rel) {
 		rel = -1;
 	}
@@ -585,7 +585,7 @@ static void mixrel_centre(VOICE v, SINT32 *dst, SAMPLE src, SAMPLE srcterm) {
 	SINT32	d;
 
 	vol = v->envleft << RELBIT;
-	rel = - (vol / (srcterm - src));
+	rel = - (vol / (int)(srcterm - src));
 	if (!rel) {
 		rel = -1;
 	}
@@ -703,7 +703,7 @@ static void mixenv_normal(VOICE v, SINT32 *dst, SAMPLE src, SAMPLE srcterm) {
 		} while(srcbreak < srcterm);
 	}
 
-	v->envcount = cnt - (srcterm - src);
+	v->envcount = cnt - (int)(srcterm - src);
 	do {
 		s = *src++;
 		dst[0] += SAMPMULVOL(s, voll);
@@ -745,7 +745,7 @@ static void mixenv_left(VOICE v, SINT32 *dst, SAMPLE src, SAMPLE srcterm) {
 		} while(srcbreak < srcterm);
 	}
 
-	v->envcount = cnt - (srcterm - src);
+	v->envcount = cnt - (int)(srcterm - src);
 	do {
 		s = *src++;
 		dst[0] += SAMPMULVOL(s, vol);
@@ -794,7 +794,7 @@ static void mixenv_centre(VOICE v, SINT32 *dst, SAMPLE src, SAMPLE srcterm) {
 		} while(srcbreak < srcterm);
 	}
 
-	v->envcount = cnt - (srcterm - src);
+	v->envcount = cnt - (int)(srcterm - src);
 	do {
 		s = *src++;
 		d = SAMPMULVOL(s, vol);
